Pdf.Document.Page

Description

PDF document page

Documentation

data Page

Pdf document page

pageParentNode :: Page -> IO PageNode

Page's parent node

pageContents :: Page -> IO [Ref]

List of references to page's content streams

pageMediaBox :: Page -> IO (Rectangle Double)

Media box, inheritable

pageFontDicts :: Page -> IO [(Name, FontDict)]

Font dictionaries for the page

pageExtractText :: Page -> IO Text

Extract text from the page

It tries to add spaces between chars if they don't present as actual characters in content stream.

pageExtractGlyphs :: Page -> IO [[Glyph]]

glyphsToText :: [[Glyph]] -> Text

Convert glyphs to text, trying to add spaces and newlines

It takes list of spans. Each span is a list of glyphs that are outputed in one shot. So we don't need to add space inside span, only between them.