Pdf.Document.PageNode

Description

Page tree node

Documentation

data PageNode

Page tree node, contains pages or other nodes

data PageTree

Page tree

Constructors

PageTreeNode PageNode

PageTreeLeaf Page

pageNodeNKids :: PageNode -> IO Int

Total number of child leaf nodes, including deep children

pageNodeParent :: PageNode -> IO (Maybe PageNode)

Parent page node

pageNodeKids :: PageNode -> IO [Ref]

Referencies to all kids

loadPageNode :: Pdf -> Ref -> IO PageTree

Load page tree node by reference

pageNodePageByNum :: PageNode -> Int -> IO Page

Find page by it's number

Note: it is not efficient for PDF files with a lot of pages, because it performs traversal through the page tree each time. Use pageNodeNKids, pageNodeKids and loadPageNode for efficient traversal.