Pdf.Document.Pdf

Documentation

data Pdf

pdfWithFile :: File -> IO Pdf

Make Pdf with interface to pdf file

pdfWithHandle :: Handle -> IO Pdf

Make Pdf with seekable handle

document :: Pdf -> IO Document

Get PDF document

lookupObject :: Pdf -> Ref -> IO Object

Find object by it's reference

streamContent :: Pdf -> Ref -> Stream -> IO (InputStream ByteString)

Get stream content, decoded and decrypted

Note: length of the content may differ from the raw one

rawStreamContent :: Pdf -> Stream -> IO (InputStream ByteString)

Get stream content without decrypting or decoding it

decryptStream :: Pdf -> Ref -> InputStream ByteString -> IO (InputStream ByteString)

Decrypt stream content

Note: length may change when decrypting. E.g. AESV2 encryption handler stored initializing vector in the first 16 bytes of the stream, and we strip them here.

decodeStream :: Pdf -> Stream -> InputStream ByteString -> IO (InputStream ByteString)

Decode stream content

It should be already decrypted

deref :: Pdf -> Object -> IO Object

Recursively load indirect object

isEncrypted :: Pdf -> IO Bool

Whether the PDF document it encrypted

setUserPassword :: Pdf -> ByteString -> IO Bool

Set the password to be user for decryption

Returns False when the password is wrong

data EncryptedError

File is enctypted

Constructors

EncryptedError Text